feat: initial support for privsep WireGuard interface management
This is still in proof-of-concept/prototype territory as not all error
handling is threaded through, the CLI parsing has not been augmented to
support some optional behavior, many panics still exist, and TODOs
remain.

This change adds IPC support for the CLI application on Unix like and
likely Windows systems (still untested). The main process spawns a child
process which will run with root permissions and performs 3 tasks:

- Writes a WireGuard configuration file for the interace
- Brings the interface up
- Brings the interface down

There is only one interface which is known to the child process so no
arbitrary interfaces can be changed. Only serializes WireGuard
configuration data structures are transmitted so this should not
introduce an arbitrary file writing escalation. Furthermore the target
file locaiton is also known to the child process so arbitrary file
locations cannot be specified. Finally, on Unixlike systems, the
privileges are dropped to a non root user after spawning the child so
the remainder of the program functions without elevated privilges.

The IPC logic exists in an `ipc` Cargo feature, the privilege
seperation/dropping is in a `privs` feature, and a function to check if
the process is running as root is in a `checkroot` feature. By default
the `application` feature includes these all, but it is possible to run
the program without IPC and privilege dropping by only selecting `clap`
for the CLI application.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>